DML DDD estimator for ATT, with panel data and 2 periods

This function implements a double machine learning estimator for assessing the average treatment effect on the treated (ATT) using a triple differences (DDD) approach in panel data settings across two time periods. The function takes preprocessed data structured specifically for this analysis.

Usage

att_dml(did_preprocessed)

Arguments

did_preprocessed

A list containing preprocessed data and specifications for the DDD estimation. Expected elements include: - preprocessed_data: A data table containing the data with variables needed for the analysis. - xformula: The formula for the covariates to be included in the model. It should be of the form ~ x1 + x2. Default is xformla = ~1 (no covariates). - learners: A list of learners to be used in the estimation. It should be a list of two elements, the first element being the learner for the propensity score and the second element being the learner for the outcome regression. Default is NULL, then OLS and MLE Logit is used to estimate nuisances parameters. If est_method = "dml", user have to specify learners. - n_folds: The number of folds to be used in the cross-fitting. Default is NULL. If est_method = "dml", user have to specify n_folds. - alpha: The level of significance for the confidence intervals. Default is 0.05. - subgroup_counts: A matrix containing the number of observations in each subgroup.

Value

A list with the estimated ATT, standard error, upper and lower confidence intervals, and n_folds.