Function that generates panel data with staggered treatment assignment and multiple periods

Function to generate data with staggered treatment adoption. Without loss of generality, the number of time periods is set to be 3.

Usage

gen_dgp_mult_periods(size, dgp_type = 1)

Arguments

size

number of units

dgp_type

type of DGP to generate. 1 if both nuisance functions are correct; 2 if only the outcome model is correct; 3 if only the pscore is correct; 4 if both nuisance functions are incorrect.

Value

A list with 2 data.table with the following columns in long and wide format.

data

data.table with the generated data in long format.

data_wide

data.table with the generated data in wide format.

id: ID for panel data.

cohort: Indicate the first period where the treatment is assigned.

partition: Partition variable.

x1: Covariate 1

x2: Covariate 2

x3: Covariate 3

x4: Covariate 4

cluster: Cluster variable (there's no actual within-cluster correlation).

time: Time periods.

y: Outcome variable.

ES_0_unfUnfeasible event-study parameter for the DGP.

prob_g2_p1Proportion of units in group 2 and eligibility = 1.

prob_g3_p1Proportion of units in group 3 and eligibility = 1.